+++ src/integrationTest/kotlin/org/jetbrains/intellij/platform/gradle/DependencyLockingAndVerificationIntegrationTest.kt
@@ -20,7 +20,6 @@ class DependencyLockingAndVerificationIntegrationTest : IntelliJPlatformIntegrat
 ) {
 
     @Test
-    @Ignore("https://github.com/JetBrains/intellij-platform-gradle-plugin/issues/1779")
     fun `build plugin with dependency locks & hash verification`() {
         build(
             Tasks.External.CLEAN,
@@ -44,7 +43,6 @@ class DependencyLockingAndVerificationIntegrationTest : IntelliJPlatformIntegrat
     }
 
     @Test
-    @Ignore("https://github.com/JetBrains/intellij-platform-gradle-plugin/issues/1779")
     fun `build plugin with dependency locks, hash & signature verification`() {
         build(
             Tasks.External.CLEAN,